Move code, and edit some __init__.py files
