+++ syne_tune/optimizer/schedulers/searchers/botorch/__init__.py
@@ -10,9 +10,6 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
-from syne_tune.optimizer.schedulers.searchers.kde.kde_searcher import (
-    KernelDensityEstimator,
-)  # noqa: F401
-from syne_tune.optimizer.schedulers.searchers.kde.multi_fidelity_kde_searcher import (
-    MultiFidelityKernelDensityEstimator,
+from syne_tune.optimizer.schedulers.searchers.botorch.botorch_searcher import (
+    BotorchSearcher,
 )  # noqa: F401
