+++ syne_tune/optimizer/schedulers/searchers/constrained/__init__.py
@@ -10,3 +10,6 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
+from syne_tune.optimizer.schedulers.searchers.constrained.constrained_gp_fifo_searcher import (
+    ConstrainedGPFIFOSearcher,
+)  # noqa: F401
