+++ syne_tune/optimizer/schedulers/searchers/hypertune/__init__.py
@@ -10,3 +10,6 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
+from syne_tune.optimizer.schedulers.searchers.hypertune.hypertune_searcher import (
+    HyperTuneSearcher,
+)  # noqa: F401
