+++ syne_tune/optimizer/schedulers/searchers/utils/__init__.py
@@ -10,3 +10,18 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
+    HyperparameterRanges,
+)  # noqa: F401
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
+    make_hyperparameter_ranges,
+)  # noqa: F401
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_impl import (
+    HyperparameterRangesImpl,
+)  # noqa: F401
+from syne_tune.optimizer.schedulers.searchers.utils.scaling import (  # noqa: F401
+    LinearScaling,
+    LogScaling,
+    ReverseLogScaling,
+    get_scaling,
+)
